refactor to reduce generated code
